openvpn@.service: Bump LimitNPROC to 100
This generally seems to be the wrong knob to protect against runaway
forks (as it does not limit per instance, but per user systemwide), but
a general mediation is still under discussion. Meanwhile bump the limit
for the Debian unit to 100.

Upstream openvpn-client@.service and openvpn-server@.service still use
10

See Bug#861923 for discussion.