Adapt autopkgtest output to 2.5
OpenVPN 2.5 slightly changed the output of some steps due to
the switch from iproute2 to the newer Netlink based interface.
This was not noticed because the autopkgtest of OpenVPN is not
run in debci infrastructure (machine-isolation not supported)

This patch was imported from Ubuntu, thanks a lot!

Partially fixes Bug#983662, but there is more work to do