d/openvpn@.service: Replace LimitNPROC=100 with TasksMax=10
There has been a longstanding issue where the default upstream
OpenVPN openvpn-client@.service and openvpn-server@.service units
set LimitNPROC=10. This is limit is not per unit instance, but
appears to be system-wide per user. As a consequence, this limit
is often way too low, and denies even a single process spawns in
an openvpn instance if other instances are misbehaving.

As a consequence, the limit was raised to 100 in the Debian
openvpn@.service in 2019, and various projects using OpenVPN
recommend or automatically install overrides for this knob (i.e.
eduVPN).

The better setting for this would be the native systemd TasksMax=
knob, which limits the maximum number of tasks that may be created
by this unit.

See but not closes Bug: #861923